Add Azure Function for Zoom webhooks that populates Azure Table Storage
Because the Zoom API doesn't allow us to get a list of users in a meeting on the free or Pro tier (you need business) - we're using Zoom's webhooks instead. This Azure Function will respond to Zoom webhook when someone enters and leaves the meeting, and add an entry in an Azure Table Storage table.