Change attendeeId from int to string and have Zoom use the uuid
The zoom userid is only unique per meeting instance, so if they leave and rejoin, it changes. So this change makes it use the user uuid instead, which doesn't change. Note that as part of this change, I've had to make the attendeeId a string across the board - for meetup aswell.